Revert editTodo to previous working version so feature is functional. Delete 'oldsingleTodo' file
